Addet tets PathToMultimedia
